cloud-run/requirements.txt
# Requirements for Sales Intelligence API v2.1
# Modular architecture with AI Analysis endpoint

fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-bigquery==3.13.0
google-cloud-firestore==2.13.1
google-cloud-aiplatform==1.73.0
google-generativeai==0.3.2
pydantic==2.4.2
python-multipart==0.0.6
numpy==1.24.3
holidays==0.66
